% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_api.R
\name{ExportFlowFrame}
\alias{ExportFlowFrame}
\title{ExportFlowFrame}
\usage{
ExportFlowFrame(ds)
}
\arguments{
\item{ds}{The dataset to be exported}
}
\value{
A new flowFrame object
}
\description{
Save the available dataset information in a flowFrame.
You can save the resulting flowFrame to FCS file using e.g.
'flowCore::write.FCS'.
}
