% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_api.R
\name{PopulationKeys}
\alias{PopulationKeys}
\title{PopulationKeys}
\usage{
PopulationKeys(ds)
}
\arguments{
\item{ds}{The dataset to be processed}
}
\value{
Named character vector with population names and keys
}
\description{
Get available population names from a dataset
}
