% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_api.R
\name{PopulationSizes}
\alias{PopulationSizes}
\title{PopulationSizes}
\usage{
PopulationSizes(ds)
}
\arguments{
\item{ds}{The dataset to be processed}
}
\value{
Table with all annotated population sizes
}
\description{
Get cell counts in annotated populations in a dataset
}
