% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_api.R
\name{Process}
\alias{Process}
\title{Process}
\usage{
Process(nds, a)
}
\arguments{
\item{nds}{Loaded cell object, e.g. as received from 'LoadCells'}

\item{a}{Loaded analysis object, e.g. as exported from ShinySOM GUI and loaded using 'readRDS'}
}
\value{
The processed dataset
}
\description{
Batch-process the cells according to the analysis saved in ds
}
