% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_api.R
\name{Dissect}
\alias{Dissect}
\title{Dissect}
\usage{
Dissect(ds, pops)
}
\arguments{
\item{ds}{the dataset to be reduced}

\item{pops}{population keys to be dissected (the resulting dataset will only
contain populations specified here)}
}
\value{
the reduced dataset object
}
\description{
Reduce the dataset to subsets, as specified by the selected populations
}
