% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_api.R
\name{LoadCells}
\alias{LoadCells}
\title{LoadCells}
\usage{
LoadCells(files, nCells = NULL, noComp = F, colsToLoad = NULL)
}
\description{
Opens several files and concatenates their contents into a dataset object.
}
